move git config parsing to git pkg
