+++ config/git_fetcher.go
@@ -16,19 +16,7 @@ type GitFetcher struct {
 	vals map[string][]string
 }
 
-type GitConfig struct {
-	Lines        []string
-	OnlySafeKeys bool
-}
-
-func NewGitConfig(gitconfiglines string, onlysafe bool) *GitConfig {
-	return &GitConfig{
-		Lines:        strings.Split(gitconfiglines, "\n"),
-		OnlySafeKeys: onlysafe,
-	}
-}
-
-func ReadGitConfig(configs ...*GitConfig) (gf *GitFetcher, extensions map[string]Extension, uniqRemotes map[string]bool) {
+func ReadGitConfig(configs ...*git.ConfigurationSource) (gf *GitFetcher, extensions map[string]Extension, uniqRemotes map[string]bool) {
 	vals := make(map[string][]string)
 	ignored := make([]string, 0)
 
@@ -161,37 +149,13 @@ func (g *GitFetcher) All() map[string][]string {
 	return newmap
 }
 
-func getGitConfigs() (sources []*GitConfig) {
-	if lfsconfig := getFileGitConfig(".lfsconfig"); lfsconfig != nil {
-		sources = append(sources, lfsconfig)
-	}
-
-	globalList, err := git.Config.List()
-	if err == nil {
-		sources = append(sources, NewGitConfig(globalList, false))
-	} else {
+func getGitConfigs() (sources []*git.ConfigurationSource) {
+	sources, err := git.Config.Sources(filepath.Join(LocalWorkingDir, ".lfsconfig"))
+	if err != nil {
 		fmt.Fprintf(os.Stderr, "Error reading git config: %s\n", err)
-	}
-
-	return
-}
-
-func getFileGitConfig(basename string) *GitConfig {
-	fullname := filepath.Join(LocalWorkingDir, basename)
-	if _, err := os.Stat(fullname); err != nil {
-		if !os.IsNotExist(err) {
-			fmt.Fprintf(os.Stderr, "Error reading %s: %s\n", basename, err)
-		}
 		return nil
 	}
-
-	lines, err := git.Config.ListFromFile(fullname)
-	if err == nil {
-		return NewGitConfig(lines, true)
-	}
-
-	fmt.Fprintf(os.Stderr, "Error reading %s: %s\n", basename, err)
-	return nil
+	return sources
 }
 
 func keyIsUnsafe(key string) bool {
