+++ git/git.go
@@ -16,7 +16,6 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	lfserrors "github.com/git-lfs/git-lfs/errors"
@@ -293,7 +292,6 @@ func RemoteBranchForLocalBranch(localBranch string) string {
 	} else {
 		return localBranch
 	}
-
 }
 
 func RemoteList() ([]string, error) {
@@ -460,133 +458,6 @@ func DefaultRemote() (string, error) {
 
 func UpdateIndexFromStdin() *subprocess.Cmd {
 	return git("update-index", "-q", "--refresh", "--stdin")
-}
-
-// Configuration can fetch or modify the current Git config and track the Git
-// version.
-type Configuration struct {
-	version string
-	mu      sync.Mutex
-}
-
-var Config = &Configuration{}
-
-// Find returns the git config value for the key
-func (c *Configuration) Find(val string) string {
-	output, _ := gitSimple("config", val)
-	return output
-}
-
-// FindGlobal returns the git config value global scope for the key
-func (c *Configuration) FindGlobal(val string) string {
-	output, _ := gitSimple("config", "--global", val)
-	return output
-}
-
-// FindSystem returns the git config value in system scope for the key
-func (c *Configuration) FindSystem(val string) string {
-	output, _ := gitSimple("config", "--system", val)
-	return output
-}
-
-// Find returns the git config value for the key
-func (c *Configuration) FindLocal(val string) string {
-	output, _ := gitSimple("config", "--local", val)
-	return output
-}
-
-// SetGlobal sets the git config value for the key in the global config
-func (c *Configuration) SetGlobal(key, val string) (string, error) {
-	return gitSimple("config", "--global", "--replace-all", key, val)
-}
-
-// SetSystem sets the git config value for the key in the system config
-func (c *Configuration) SetSystem(key, val string) (string, error) {
-	return gitSimple("config", "--system", "--replace-all", key, val)
-}
-
-// UnsetGlobal removes the git config value for the key from the global config
-func (c *Configuration) UnsetGlobal(key string) (string, error) {
-	return gitSimple("config", "--global", "--unset", key)
-}
-
-// UnsetSystem removes the git config value for the key from the system config
-func (c *Configuration) UnsetSystem(key string) (string, error) {
-	return gitSimple("config", "--system", "--unset", key)
-}
-
-// UnsetGlobalSection removes the entire named section from the global config
-func (c *Configuration) UnsetGlobalSection(key string) (string, error) {
-	return gitSimple("config", "--global", "--remove-section", key)
-}
-
-// UnsetSystemSection removes the entire named section from the system config
-func (c *Configuration) UnsetSystemSection(key string) (string, error) {
-	return gitSimple("config", "--system", "--remove-section", key)
-}
-
-// UnsetLocalSection removes the entire named section from the system config
-func (c *Configuration) UnsetLocalSection(key string) (string, error) {
-	return gitSimple("config", "--local", "--remove-section", key)
-}
-
-// SetLocal sets the git config value for the key in the specified config file
-func (c *Configuration) SetLocal(file, key, val string) (string, error) {
-	args := make([]string, 1, 6)
-	args[0] = "config"
-	if len(file) > 0 {
-		args = append(args, "--file", file)
-	}
-	args = append(args, "--replace-all", key, val)
-	return gitSimple(args...)
-}
-
-// UnsetLocalKey removes the git config value for the key from the specified config file
-func (c *Configuration) UnsetLocalKey(file, key string) (string, error) {
-	args := make([]string, 1, 5)
-	args[0] = "config"
-	if len(file) > 0 {
-		args = append(args, "--file", file)
-	}
-	args = append(args, "--unset", key)
-	return gitSimple(args...)
-}
-
-// List lists all of the git config values
-func (c *Configuration) List() (string, error) {
-	return gitSimple("config", "-l")
-}
-
-// ListFromFile lists all of the git config values in the given config file
-func (c *Configuration) ListFromFile(f string) (string, error) {
-	return gitSimple("config", "-l", "-f", f)
-}
-
-// Version returns the git version
-func (c *Configuration) Version() (string, error) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	if len(c.version) == 0 {
-		v, err := gitSimple("version")
-		if err != nil {
-			return v, err
-		}
-		c.version = v
-	}
-
-	return c.version, nil
-}
-
-// IsVersionAtLeast returns whether the git version is the one specified or higher
-// argument is plain version string separated by '.' e.g. "2.3.1" but can omit minor/patch
-func (c *Configuration) IsGitVersionAtLeast(ver string) bool {
-	gitver, err := c.Version()
-	if err != nil {
-		tracerx.Printf("Error getting git version: %v", err)
-		return false
-	}
-	return IsVersionAtLeast(gitver, ver)
 }
 
 // RecentBranches returns branches with commit dates on or after the given date/time
